fix delete queue

+++ app/code/local/Citrus/Integration/Model/Resource/Queue.php
@@ -14,9 +14,11 @@ class Citrus_Integration_Model_Resource_Queue extends Mage_Core_Model_Resource_D
     public function makeDelete($ids, $type = null)
     {
         $adapter = $this->_getReadAdapter();
-        if($type)
-            $adapter->delete(self::getMainTable(), 'id in ('.implode(',',$ids).') and type = "'.$type.'"');
-        else
-            $adapter->delete(self::getMainTable(), 'id in ('.implode(',',$ids).')');
+        if ($ids) {
+            if ($type)
+                $adapter->delete(self::getMainTable(), 'id in (' . implode(',', $ids) . ') and type = "' . $type . '"');
+            else
+                $adapter->delete(self::getMainTable(), 'id in (' . implode(',', $ids) . ')');
+        }
     }
 }
